Refactoring and Ekstend lib introduction

+++ library/src/main/java/de/troido/bleacon/ble/Nordic.kt
@@ -1,8 +1,9 @@
 package de.troido.bleacon.ble
 
-import de.troido.bleacon.util.toUInt
+import de.troido.ekstend.numeric.toUShort
 
 /** Nordic Semiconductors manufacturer ID. */
 val NORDIC_ID = toManufacturerId(byteArrayOf(0x59, 0x00))
 
-private fun toManufacturerId(data: ByteArray): Int = (data[1].toUInt() shl 8) + data[0].toUInt()
+private fun toManufacturerId(data: ByteArray): Int =
+        (data[1].toUShort() shl 8) + data[0].toUShort()
